% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDensity.R
\name{cal.motif.dens}
\alias{cal.motif.dens}
\title{Calculate motif density of input bed file}
\arguments{
\item{fasta}{Fasta file converted from input file by 'bedtools getfasta'}

\item{bed}{Input bed file}

\item{motif}{Input motif you want to calculate the density}
}
\value{
Bed file added the column of calculated input motif density
}
\description{
\code{cal.motif.dens} cal.motif.dens
}
